populate "judge" fields based on combination of encoding vs decoding and fixed vs variable.
